Created file path handling lib + Some work in install/play callback.
Created a path handling library which will in charge of all path
manipulation functionalities, and some unittests for it.

Some work in install/play callback. At the moment it just performs the
basic functionality and the installation process is still missing.

Added some basic "logging" of options chosen in the menu system which
are printed to the terminal.