Added patch dir indexing function + test data dir standardization
Added a function to index patch dirs, so the index matches ROM files
indices and a patch dir can be directly applied over a rom dir and every
fill will have the right patch applied.

Created a function to standardise the input directory for unit tests. At
the moment, only the patch dir index function is using the standard
input directory and further work is needed for other unit tests.